chore: rename component files to PascalCase (layout/ + rivalhub/)
