+++ src/app/[seasonSlug]/layout.tsx
@@ -4,8 +4,8 @@ import type { Metadata } from "next";
 import { eq, and, count } from "drizzle-orm";
 import { db } from "@/db/client";
 import { seasons, seasonRegistrations } from "@/db/schema";
-import { Breadcrumb } from "@/components/layout/breadcrumb";
-import { SeasonNav } from "@/components/layout/season-nav";
+import { Breadcrumb } from "@/components/layout/Breadcrumb";
+import { SeasonNav } from "@/components/layout/SeasonNav";
 import { hexToRgbString } from "@/lib/utils/color";
 import { normalizeStagePlan } from "@/types/season";
 import { showStats } from "@/lib/utils/season";
