+++ src/components/draft/CaptainDraftPanel.tsx
@@ -9,8 +9,8 @@ import { Button } from "@/components/ui/button";
 import { DraftCountdown } from "./DraftCountdown";
 import { canPickPosition } from "@/lib/draft/rules";
 import { positionLabel } from "@/lib/validators/registration";
-import { MapPreferenceChips } from "@/components/rivalhub/map-preference-chips";
-import { PosChip } from "@/components/rivalhub/pos-chip";
+import { MapPreferenceChips } from "@/components/rivalhub/MapPreferenceChips";
+import { PosChip } from "@/components/rivalhub/PosChip";
 import { PlayerInfoPopover } from "./PlayerInfoPopover";
 import { getDisplayName } from "@/lib/utils/display-name";
 import { sortByRank } from "@/lib/utils/rank";
