+++ src/components/draft/PlayerPool.tsx
@@ -5,8 +5,8 @@ import { useMemo, useState } from "react";
 import Link from "next/link";
 import type { DraftPlayerRow } from "@/lib/draft/data";
 import { positionLabel } from "@/lib/validators/registration";
-import { MapPreferenceChips } from "@/components/rivalhub/map-preference-chips";
-import { PosChip } from "@/components/rivalhub/pos-chip";
+import { MapPreferenceChips } from "@/components/rivalhub/MapPreferenceChips";
+import { PosChip } from "@/components/rivalhub/PosChip";
 import { PlayerInfoPopover } from "./PlayerInfoPopover";
 import { getDisplayName } from "@/lib/utils/display-name";
 import { sortByRank } from "@/lib/utils/rank";
