+++ src/components/rivalhub/EmptyState.tsx
