+++ src/components/rivalhub/MapPreferenceChips.tsx
