+++ src/components/rivalhub/Panel.tsx
