+++ src/components/rivalhub/ScrollHint.tsx
