+++ src/components/rivalhub/Stat.tsx
