+++ src/components/rivalhub/StatusBanner.tsx
