+++ src/components/rivalhub/StatusPill.tsx
