+++ src/components/rivalhub/TeamBadge.tsx
