+++ src/components/rivalhub/index.ts
@@ -1,15 +1,15 @@
-export { Panel } from "./panel";
-export { Btn } from "./btn";
-export { Stat, MiniStat } from "./stat";
-export { Marker } from "./marker";
-export { Field } from "./field";
-export { StatusBanner } from "./status-banner";
-export { InlineConfirm } from "./inline-confirm";
-export { EmptyState } from "./empty-state";
-export { ErrorState } from "./error-state";
-export { RivalSkeleton, SkeletonRow, Spinner } from "./skeleton";
-export { TeamBadge } from "./team-badge";
-export { PosChip } from "./pos-chip";
-export { StatusPill } from "./status-pill";
-export { ScrollHint } from "./scroll-hint";
-export { PhaseStep } from "./phase-step";
+export { Panel } from "./Panel";
+export { Btn } from "./Btn";
+export { Stat, MiniStat } from "./Stat";
+export { Marker } from "./Marker";
+export { Field } from "./Field";
+export { StatusBanner } from "./StatusBanner";
+export { InlineConfirm } from "./InlineConfirm";
+export { EmptyState } from "./EmptyState";
+export { ErrorState } from "./ErrorState";
+export { RivalSkeleton, SkeletonRow, Spinner } from "./Skeleton";
+export { TeamBadge } from "./TeamBadge";
+export { PosChip } from "./PosChip";
+export { StatusPill } from "./StatusPill";
+export { ScrollHint } from "./ScrollHint";
+export { PhaseStep } from "./PhaseStep";
